private/src/requirements.txt
Flask 
flask-cors 
psutil
fastapi
starlette
uvicorn
python-dotenv